cgosymbolizer: add context callback for runtime.SetCgoTraceback
In this implementation, the context callback collects as much of the
call stack as it can, and that stack is subsequently copied in the
traceback callback. According to the runtime.SetCgoTraceback
documentation, this behavior is correct albeit (possibly) inefficient.

However, it is less clear to me if the implementation in this commit is
completely correct as far as using a thread-local call stack data
structure is concerned. Specifically, the "traceback function may be
called from an execution thread other than the one that recorded the
context, but only when the context is known to be valid and unchanging".
I read this to mean that the other execution thread will not have called
the context callback again before it's used. But it's not 100% clear
that that is the case.